renamed "query" to "snappySql"
- renamed query implicit methods to snappySql as per comments elsewhere
- split JdbcReader into separate file
